backend/marketing_os_app/requirements.txt
pytest==8.0.0
pytest-asyncio==0.23.5
sqlalchemy==2.0.25
pydantic==2.6.1
pyjwt==2.8.0
fastapi==0.109.2
